+++ CHANGELOG.md
@@ -115,7 +115,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added Tony to 'about us'
 - Added Igor to 'about us'
 - Added GPL v3 license
-- Updated robots.txt and SEO tags
+- Added netlify icon and text to the footer
+- Updated robots.txt and added SEO tags
 
 ### Fixed